updated test cases
changed a bit of code from DbBookFunctions.java and DbConnector.java for the test cases to be valid, since static objects cannot be changed in runtime. Had to change functions and their implementation so that it all runs smoothly now. All test cases work and Code coverage is on high coverage now.